+++ prac_06/languages.py
@@ -1,24 +1,23 @@
 """
-Cp1404 Practical 6
-Create programming language class
+CP1404 Practical 6
+Programming languages
+Estimated time: 20 minutes
+Actual time: 25 minutes
 """
 
+from prac_06.programming_language import ProgrammingLanguage
 
-class ProgrammingLanguage:
-    """Represent a Programming Language object."""
+python = ProgrammingLanguage("Python", "Dynamic", True, 1991)
+ruby = ProgrammingLanguage("Ruby", "Dynamic", True, 1995)
+visual_basic = ProgrammingLanguage("Visual Basic", "Static", False, 1991)
+print(python)
 
-    def __init__(self, name="", typing=bool, reflection="", year=0):
-        """Initialise a Programming Language instance. """
-        self.name = name
-        self.typing = typing
-        self.reflection = reflection
-        self.year = year
+languages = [python, ruby, visual_basic]
+print("The dynamically typed languages are:")
+for language in languages:
+    if language.is_dynamic():
+        print(language.name)
 
-    def __str__(self):
-        """Return data as a formatted f string."""
-        return f"{self.name}, {self.typing} Typing, Reflection={self.reflection}, First appeared in {self.year}"
 
-    def is_dynamic(self):
-        """Determine if language typing is dynamic."""
-        if self.typing == "Dynamic":
-            return True
+
+
