+++ prac_06/programming_language.py
@@ -1,23 +1,24 @@
 """
-CP1404 Practical 6
-Programming languages
-Estimated time: 20 minutes
-Actual time: 25 minutes
+Cp1404 Practical 6
+Create programming language class
 """
 
-from prac_06.languages import ProgrammingLanguage
 
-python = ProgrammingLanguage("Python", "Dynamic", True, 1991)
-ruby = ProgrammingLanguage("Ruby", "Dynamic", True, 1995)
-visual_basic = ProgrammingLanguage("Visual Basic", "Static", False, 1991)
-print(python)
+class ProgrammingLanguage:
+    """Represent a Programming Language object."""
 
-languages = [python, ruby, visual_basic]
-print("The dynamically typed languages are:")
-for language in languages:
-    if language.is_dynamic():
-        print(language.name)
+    def __init__(self, name="", typing=bool, reflection="", year=0):
+        """Initialise a Programming Language instance. """
+        self.name = name
+        self.typing = typing
+        self.reflection = reflection
+        self.year = year
 
+    def __str__(self):
+        """Return data as a formatted f string."""
+        return f"{self.name}, {self.typing} Typing, Reflection={self.reflection}, First appeared in {self.year}"
 
-
-
+    def is_dynamic(self):
+        """Determine if language typing is dynamic."""
+        if self.typing == "Dynamic":
+            return True
